fix(server): Correctly wait for server.listen event (#589)
* fix(server): Add @types for server dependencies & fix app.listen calls

This adds `@types` packages for Koa modules and shortid, strengthening 
the typing of the server.

It also changes how Koa’s `app.listen` method is used. Previously, the 
server awaited `app.listen` and “then” logged the server port to the 
console. Except `listen` doesn’t return a promise, so this was 
potentially buggy. I’ve never had an issue with this but, as reported in 
#584, it seems it could create an error if `server.address()` returned 
`null` (presumably because it was called too early).

Closes #584.

* fix(server): Wait for app.listen calls & fix listen mock in test

* refactor(server): Add types for server options

* test(server): Remove nonexistent “singlePort” option from test

* refactor(api): Tidy up credentials typing

* add external to server rollup config

Co-authored-by: Nicolo Davis <nicolodavis@gmail.com>