feat: Centralise lobby API and master authentication logic (#853)
* refactor(transport): Factor out copy-paste to shared method

* refactor(transport): Factor out copy-paste to shared method

* feat(transport): Track credentials in transport class

* refactor(server): Move `IsSynchronous` type guard to `db/base`

* fix(client): Pass credentials when initialising transport

* refactor(transport): Improve typing in socket transport

* feat(auth): Encapsulate current authentication logic in new Auth class

* refactor(master): Use Auth class instead of internal logic

* feat(types): Add custom Koa context and application types

* refactor(server): Import socket.io types via top-level namespace

Helps avoid naming collisions.

* refactor(server): Use custom Koa/server types

* feat(server): Pass Auth class from transport to master

* feat(server): Use Auth class to authenticate Lobby API

Closes #693

* feat(server): Initialise Auth at top-level of server

Create the central Auth instance, add it to the app context and pass it 
when creating the router.

* feat(server): Include auth & db on API app context for consistency

* test(server): Fix types & make auth available in lobby API tests

* feat(server): Track credentials in clientInfo

* feat(master): Add credentials/authentication to all methods

Closes #840

* feat(server): Update socket transport for new master arguments

* test(server): Make auth available in transport tests

* feat(client): Update socket.io transport for new master arguments

* test(client): Fix socket transport tests

* test(server): Fix socket transport tests

Refactors the mock socket implementation to correctly simulate multiple 
open sockets.

* refactor(master): Don’t use sync storage == sync auth heuristic

Previously `authenticateCredentials` was awaited if the master was using 
an async storage implementation, but not when the master was using 
synchronous storage. This was so as to avoid `await` when running the 
master locally where it can be treated as a synchronous API. This 
assumption could be buggy though if a user provides an async 
`authenticateCredentials` method (e.g. to call out to a third-party auth 
service), but is using the default synchronous `InMemory` storage.

This commit refactors the authentication logic to assume that if the 
master has been provided with an `Auth` instance, it can be run 
asynchronously. Local masters will run without an `Auth` instance, so 
will skip authentication entirely.

* refactor(master): Add explicit return types to onUpdate and onSync

* test(master): Update auth tests and test onSync & onConnectionChange

* refactor(auth): Simplify `doesMatchRequireAuthentication` logic